Fix foreign key issue and refactor db models
